+++ crates/ibc/src/core/ics04_channel/commitment.rs
@@ -1,9 +1,10 @@
 //! Types and utilities related to packet commitments.
 
-use crate::core::ics04_channel::msgs::acknowledgement::Acknowledgement;
 use crate::core::ics04_channel::timeout::TimeoutHeight;
 use crate::core::timestamp::Timestamp;
 use crate::prelude::*;
+
+use super::packet::Acknowledgement;
 
 /// Packet commitment
 #[cfg_attr(
