+++ crates/ibc/src/mock/context.rs
@@ -1466,8 +1466,7 @@ mod tests {
 
     use crate::core::ics04_channel::channel::{Counterparty, Order};
     use crate::core::ics04_channel::error::ChannelError;
-    use crate::core::ics04_channel::msgs::acknowledgement::Acknowledgement;
-    use crate::core::ics04_channel::packet::Packet;
+    use crate::core::ics04_channel::packet::{Acknowledgement, Packet};
     use crate::core::ics04_channel::Version;
     use crate::core::ics24_host::identifier::ChainId;
     use crate::core::ics24_host::identifier::{ChannelId, ConnectionId, PortId};
